Add module .transport.check
- Static and dynamic collections of checks of model build data
  previously in .tests.model.transport.test_build.
- Simplify .test_build.test_debug().
